revert changes during trial
