+++ assets/js/index.js
@@ -1,9 +1,10 @@
+// import chatForm from "./chat.js";
 import RegisterForm from "./reg.js";
 import loginForm from "./login.js";
 import PostForm from "./post.js";
 import logoutBtn from "./logout.js";
-// import chatForm from "./chat.js";
-// const chatBox = document.querySelector(".col-1")
+import chatForm from "./chat.js";
+const chatBox = document.querySelector(".col-1")
 const loginArea = document.querySelector("#userPopUpPOne")
 const loginInputs = loginArea.firstElementChild
 const regArea = document.querySelector("#userPopUpPTwo")
